# publisher/docs/swagger.yaml
basePath: /
definitions:
  controllers.sent:
    properties:
      message:
        type: string
    type: object
...
  model.EmailList:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      deletedAt:
        type: string
      deletedBy:
        type: string
      description:
        type: string
      emailAlias:
        type: string
      emails:
        additionalProperties:
          type: string
        description: Email uses map[id] to have the possibility to remove by ID without
          revealing which email will be removed
        type: object
      id:
        type: string
      name:
        type: string
    type: object
  model.EmailListEmails:
    properties:
      emails:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - emails
    type: object
  model.EmailListInfo:
    properties:
      description:
        type: string
      emailAlias:
        type: string
      name:
        type: string
    required:
    - description
    - emailAlias
    - name
    type: object
  model.EmailListPartial:
    properties:
      description:
        type: string
      emailAlias:
        type: string
      emails:
        items:
          type: string
        minItems: 1
        type: array
      name:
        type: string
    required:
    - description
    - emailAlias
    - emails
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is an api that publishes emails in RabbitMQ.
  title: Publisher Emails
  version: "1.0"
paths:
...
  /email/list:
    get:
      consumes:
      - application/json
      description: Get all user email list.
      produces:
      - application/json
      responses:
        "200":
          description: email list
          schema:
            items:
              $ref: '#/definitions/model.EmailList'
            type: array
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get all user email list
      tags:
      - emailList
    post:
      consumes:
      - application/json
      description: Create a email list to user.
      parameters:
      - description: email list params
        in: body
        name: emailList
        required: true
        schema:
          $ref: '#/definitions/model.EmailListPartial'
      produces:
      - application/json
      responses:
        "201":
          description: create email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email list param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: email list already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Creating email list
      tags:
      - emailList
  /email/list/{name}:
    delete:
      consumes:
      - application/json
      description: Delete an email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: delete email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Delete email list
      tags:
      - emailList
    get:
      consumes:
      - application/json
      description: Get a user email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: email list
          schema:
            $ref: '#/definitions/model.EmailList'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list not found
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Get a user email list
      tags:
      - emailList
    put:
      consumes:
      - application/json
      description: Update information for an email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: email list info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/model.EmailListInfo'
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email list param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "409":
          description: email list already exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Update info email list
      tags:
      - emailList
  /email/list/{name}/add:
    post:
      consumes:
      - application/json
      description: Add emails to email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: emails
        in: body
        name: emails
        required: true
        schema:
          $ref: '#/definitions/model.EmailListEmails'
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Add emails to email list
      tags:
      - emailList
  /email/list/{name}/remove:
    delete:
      consumes:
      - application/json
      description: Remove emails to email list.
      parameters:
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: emails
        in: body
        name: emails
        required: true
        schema:
          $ref: '#/definitions/model.EmailListEmails'
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "400":
          description: an invalid email param was sent
          schema:
            $ref: '#/definitions/controllers.sent'
        "401":
          description: user session has expired
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Remove emails to email list
      tags:
      - emailList
  /email/list/{user_id}/{name}/{email_id}:
    delete:
      consumes:
      - application/json
      description: Remove emails to email list.
      parameters:
      - description: user id from emailist
        in: path
        name: user_id
        required: true
        type: string
      - description: email list name
        in: path
        name: name
        required: true
        type: string
      - description: email id from emailist
        in: path
        name: email_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: update email list successfully
          schema:
            $ref: '#/definitions/controllers.sent'
        "404":
          description: email list does not exist
          schema:
            $ref: '#/definitions/controllers.sent'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/controllers.sent'
      summary: Remove email to email list
      tags:
      - emailList
...
swagger: "2.0"